Rename webHookHandler service to webhook
