+++ src/routes/webhook.ts
@@ -1,11 +1,11 @@
 import { Request, Response } from "express";
 import logger from "../services/logger";
-import webHookHandler from "../services/webHookHandler";
+import webhook from "../services/webhook";
 
 // Listener for GitLab WebHooks
 export default async ({ body: payload, headers }: Request, res: Response) => {
   try {
-    await webHookHandler.handle({headers, payload});
+    await webhook.handle({headers, payload});
     res.send("ok\n");
   } catch (e: any) {
     logger.error(e.message);
